+++ controller.py
@@ -6,6 +6,7 @@ import traceback
 
 
 from StreamDeck.Devices.StreamDeck import StreamDeck
+from pages import get_page, MainPage
 
 LOGGER = logging.getLogger(__name__)
 
@@ -19,29 +20,13 @@ class Controller:
 
         self._lock = asyncio.Lock(loop=self.loop)
 
-        page = self.load_page("main")
+        page = MainPage(self)
         self.default_page = page
         self.current_page = page
         self.previous_page = None
-        self.page_cache["main"] = page
-
-        self.force_reload_page = False
-        self.heartbeat_delay = 60
+        self.page_cache["MainPage"] = page
 
         self.current_heartbeat_task = None
-
-
-    @staticmethod
-    def load_page(page):
-        LOGGER.info(f"Loading page {page}")
-        path = pathlib.Path("pages", f"{page}.py")
-        if path.exists():
-            ns = {"__name__": __name__}
-            exec(path.read_bytes(), ns)
-            if "PAGE" in ns:
-                return ns["PAGE"]
-        LOGGER.warning(f"Page {page} not found")
-        return None
 
     async def set_next_page(self, page):
         if page in self.page_cache:
@@ -49,7 +34,7 @@ class Controller:
             self.current_page = self.page_cache[page]
             return None
         
-        if (new_page:=self.load_page(page)) is not None:
+        if (new_page:=await get_page(page)) is not None:
             self.current_page = self.page_cache[page] = new_page
             return None
         
@@ -68,8 +53,7 @@ class Controller:
     async def update_deck(self, deck=None):
         deck = deck or self.deck
         LOGGER.info(f"Updating deck {deck.id()}")
-        await self.current_page.draw_buttons(deck)
-
+        await self.current_page.render()
 
     async def set_image(self, button: int, image):
         """
@@ -78,23 +62,18 @@ class Controller:
         async with self._lock:
             self.deck.set_key_image(button, image)
 
-
-
     async def setup(self, deck):
-        await self.current_page.setup(self)
+        await self.current_page.setup()
 
     def heartbeat(self):
-        self.loop.create_task(self.current_page.heartbeat, name="Page-heartbeat")
-        self.loop.call_later(self.heartbeat_delay, self.heartbeat)
+        pass
 
     async def __call__(self, deck, key, state):
-        await self.current_page.action(deck, key, state, self)
-        if self.force_reload_page:
-            LOGGER.info("Force reloading page")
-        
-        if (not state) or self.force_reload_page:
-            LOGGER.info(f"Button released, updating deck")
+        LOGGER.info(f"Deck {deck.id()} button {key} {'pressed' if state else 'released'}" )
+        await self.current_page.dispatch(key, state)
+
+        if not state:
             await self.update_deck(deck)
-            self.force_reload_page = False
+
