+++ pages/main.py
@@ -1,48 +1,85 @@
+import asyncio
 import logging
+
 import simpleobsws as obs
 
-from pages.buttons import (
-    Button, CommandButton, MenuButton, StatefulButton
-)
-from pages.page import Page
+from .base import Page, create_action_method
+from .commands import launch_shell
 
 LOGGER = logging.getLogger(__name__)
 
-class TerminalButton(CommandButton):
-    icon = "terminal-code.png"
-    text = "Terminal"
-    command = "gnome-terminal"
 
+class MainPage(Page):
+    """
+    Main page for my StreamDeck
+    """
 
-class ClockButton(Button):
-    icon = "clock.png"
-    text = "Clock"
+    deck_type = "StreamDeckMini"
 
-
-class MicButton(Button):
-    icon = "mic-on.png"
-    text = "Microphone"
-    
-
-class MainMenuButton(MenuButton):
-    icon = "navigation.png"
-    text = "Menu"
-    menu_name = "main-menu"
-
-
-class PlayRecordButton(StatefulButton):
-    states = {
-        "stopped": ("play-record.png", "Play/Record"),
-        "recording": ("play-pause.png", "Pause"),
-        "paused": ("play-record.png", "Resume")
+    button_1_label = "Terminal"
+    button_2_label = "Clock"
+    button_3_label = "Mute"
+    button_4_label = "Menu"
+    button_5_label = {
+        "stopped": "Start/Record",
+        "recording": "Pause",
+        "paused": "Resume"
     }
-    default_state = "stopped"
+    button_6_label = "Stop"
 
-    async def setup(self, controller, page):
-        self.page = page
+    button_1_icon = "terminal-code.png"
+    button_2_icon = "clock.png"
+    button_3_icon = "mic-on.png"
+    button_4_icon = "navigation.png"
+    button_5_icon = {
+        "stopped": "play-record.png",
+        "recording": "play-pause.png",
+        "paused": "play-record.png"        
+    }
+    button_6_icon = "play-stop.png"
+
+    def __init__(self, controller):
+        super().__init__(controller)
+
+        self.obs_state = "stopped"
+        self.obs_ws = None
+        
+    async def render(self):
+
+        async with self._lock:
+            obs_state = self.obs_state
+
+        button_1_label = self.button_1_label
+        button_2_label = self.button_2_label
+        button_3_label = self.button_3_label
+        button_4_label = self.button_4_label
+        button_5_label = self.button_5_label[obs_state]
+        button_6_label = self.button_6_label
+
+        button_1_icon = self.button_1_icon
+        button_2_icon = self.button_2_icon
+        button_3_icon = self.button_3_icon
+        button_4_icon = self.button_4_icon
+        button_5_icon = self.button_5_icon[obs_state]
+        button_6_icon = self.button_6_icon
+
+        images = await asyncio.gather(
+            self.render_image_from_file(button_1_icon, button_1_label),
+            self.render_image_from_file(button_2_icon, button_2_label),
+            self.render_image_from_file(button_3_icon, button_3_label),
+            self.render_image_from_file(button_4_icon, button_4_label),
+            self.render_image_from_file(button_5_icon, button_5_label),
+            self.render_image_from_file(button_6_icon, button_6_label)
+        )
+        
+        await asyncio.gather(
+            *[self.controller.set_image(i, image)
+            for i, image in enumerate(images)]
+        )
+
 
     async def obs_call(self, cmd, data=None):
-        ws = self.page.obs_ws
+        ws = self.obs_ws
         try:
             result = await ws.call(cmd, data)
         except obs.ConnectionFailure:
@@ -54,84 +91,11 @@ class PlayRecordButton(StatefulButton):
         else:
             if result["status"] == "error":
                 LOGGER.error(f"OBS error: {result['error']}")
-          
-
-    async def action(self, deck, state, controller, page):
-        LOGGER.info("Calling Play/Pause action")
-        if not state:
-            return await super().action(deck, state, controller, page)
-        
-        if not await self.page.obs_conn_alive():
-            connected = await self.page.connect_obs()
-            if not connected:
-                LOGGER.info("No OBS connection, returning")
-                return await super().action(deck, state, controller, page)
-
-        if self.state == "stopped":
-            LOGGER.info("Starting OBS recording")
-            await self.obs_call("StartRecording")
-        elif self.state == "paused":
-            LOGGER.info("Resuming OBS recording")
-            await self.obs_call("ResumeRecording")
-        elif self.state == "recording":
-            LOGGER.info("Pausing OBS recording")
-            await self.obs_call("PauseRecording")
-        return await super().action(deck, state, controller, page)
-
-
-class PlayStopButton(Button):
-    icon = "play-stop.png"
-    text = "Stop"
-
-    async def setup(self, controller, page):
-        self.page = page
-
-    async def obs_call(self, cmd, data=None):
-        ws = self.page.obs_ws
-        try:
-            await ws.call(cmd, data)
-        except obs.ConnectionFailure:
-            LOGGER.error("Connection to OBS failed")
-        except obs.MessageTimeout:
-            LOGGER.error("Connection to OBS timed out")
-        except obs.MessageFormatError:
-            LOGGER.error("Incorrect OBS message format")
-
-
-    async def action(self, deck, state, controller, page):
-        if not state:
-            return await super().action(deck, state, controller, self)
-        
-        await self.obs_call("StopRecording")
-        return await super().action(deck, state, controller, page)
-
-
-
-class MainPage(Page):
-
-    async def connection_lost_callback(self, data=None):
-        LOGGER.info("OBS connection lost event callback")
-        await self.obs_ws.disconnect()
-        self.buttons[4].state = self.buttons[4].default_state
-        await self.controller.update_deck()
-        self.obs_ws = None
-
-    async def recording_started_callback(self, data=None):
-        LOGGER.info("Recording started event callback")
-        self.buttons[4].state = "recording"
-        await self.controller.update_deck()
-
-    async def recording_paused_callback(self, data=None):
-        LOGGER.info("Recording paused event callback")
-        self.buttons[4].state = "paused"
-        await self.controller.update_deck()
-
-    async def recording_stopped_callback(self, data=None):
-        LOGGER.info("Recording stopped event callback")
-        self.buttons[4].state = "stopped"
-        await self.controller.update_deck()
 
     async def connect_obs(self):
+        """
+        Connect to the OBS web socket
+        """
         try:
             await self.obs_ws.connect()
             LOGGER.info("Connected to OBS")
@@ -139,33 +103,92 @@ class MainPage(Page):
         except ConnectionError:
             LOGGER.info("Connection to OBS failed")
         return False
-        
-
-    async def setup(self, controller):
-        await super().setup(controller)
-
-        self.obs_ws = ws = obs.obsws(loop=controller.loop)
-        ws.register(self.connection_lost_callback, "Exiting")
-        ws.register(self.recording_started_callback, "RecordingStarted")
-        ws.register(self.recording_paused_callback, "RecordingPaused")
-        ws.register(self.recording_stopped_callback, "RecordingStopping")
-        await self.connect_obs()
 
     async def obs_conn_alive(self):
         return self.obs_ws.recv_task is not None
 
-    async def heartbeat(self, deck, controller):
-        LOGGER.info("Running Main page heartbeat")
-        await self.connect_obs()
-        await super().heartbeat(deck, controller)
+    async def connection_lost_callback(self, data=None):
+        LOGGER.info("OBS connection lost event callback")
+        async with self._lock:
+            await self.obs_ws.disconnect()
+            self.obs_state = "stopped"
+            await self.controller.update_deck()
+            self.obs_ws = None
 
+    async def recording_started_callback(self, data=None):
+        async with self._lock:
+            LOGGER.info("Recording started event callback")
+            self.obs_state.state = "recording"
+            await self.controller.update_deck()
 
-PAGE = MainPage(
-    "Main",
-    TerminalButton,
-    ClockButton,
-    MicButton,
-    MainMenuButton,
-    PlayRecordButton,
-    PlayStopButton
-)
+    async def recording_paused_callback(self, data=None):
+        async with self._lock:
+            LOGGER.info("Recording paused event callback")
+            self.obs_state = "paused"
+            await self.controller.update_deck()
+
+    async def recording_stopped_callback(self, data=None):
+        async with self._lock:
+            LOGGER.info("Recording stopped event callback")
+            self.obs_state = "stopped"
+            await self.controller.update_deck()
+
+    button_1 = create_action_method(launch_shell, "gnome-terminal")
+    alt_button_1 = button_1
+
+    async def button_2(self):
+        """
+        Display clock page
+        """
+        pass
+
+    alt_button_2 = button_2
+
+    async def button_3(self):
+        """
+        Mute microphone
+        """
+        pass
+
+    alt_button_3 = button_3
+
+    async def button_4(self):
+        """
+        Open main menu page
+        """
+        pass
+
+    alt_button_4 = button_4
+
+    async def button_5(self):
+        """
+        Start/pause recording
+        """
+        
+        if not await self.obs_conn_alive():
+            if not await self.connect_obs():
+                LOGGER.info("No OBS connection, returning")
+                return None
+
+        with self._lock:
+            state = self.obs_state
+
+        if state == "stopped":
+            LOGGER.info("Starting OBS recording")
+            await self.obs_call("StartRecording")
+        elif state == "paused":
+            LOGGER.info("Resuming OBS recording")
+            await self.obs_call("ResumeRecording")
+        elif state == "recording":
+            LOGGER.info("Pausing OBS recording")
+            await self.obs_call("PauseRecording")
+
+    alt_button_5 = button_5
+
+    async def button_6(self):
+        """
+        Stop recording
+        """
+        await self.obs_call("StopRecording")
+
+    alt_button_6 = button_6
